Load checks from a registry
